+++ Sources/Sentry/SentryReplayApi.m
@@ -73,6 +73,18 @@
     [replayIntegration stop];
 }
 
+- (void)showMaskPreview {
+    [self showMaskPreview:1];
+}
+
+- (void)showMaskPreview:(CGFloat)opacity {
+    SentrySessionReplayIntegration *replayIntegration
+        = (SentrySessionReplayIntegration *)[SentrySDK.currentHub
+            getInstalledIntegration:SentrySessionReplayIntegration.class];
+    
+    [replayIntegration showMaskPreview:opacity];
+}
+
 @end
 
 #endif